Day 2 updates
I initially felt very overwhelmed so I didn't really knowing where to go but I started on the cleaning_data function which I started out approaching it all wrong but when I noticed what I was actually doing, I finally got the height to display in integers and the experience level to display a boolean and even got one of the extra credits done, the guardian list. I don't know if this what was implied or not but I zipped each guardian with the corresponding child because not all the kids had the same last name as their guardian so, if I wanted to actually be able to use that info later, I made a list to where if you enter the kid's name, you'll get the corresponding guardian name. I also started on the experience level portion of the extra credit, I have them separated into lists. With the exp/inexp players, I plan to just append 3 from each list to a team, which should give me the correct amount on each team with half exp, half inexp.