Another stepping stone
In this version, I was trying to figure out how to exit upon request. I toyed around with using sys.exit but I didn't like the output traceback and couldn't figure out how to get it to not display that so I ended up scrapping the idea along with fixing some stuff before ultimately changing and fixing a lot of bugs.