Final Update
This is mostly my final update, if nothing else needs to be fixed. I struggled in the beginning with this project, mainly on where to get started, but I trudged through! Between the works in progress and this final version, I changed the sys.exit to a while loop; I added more lists, esp. the ones to iterate through to print, such as the player names and the ones to validate; and changing the validation steps and outputs. I think if I summed up the struggles, it would be trying to remember and use all the rules behind lists and dictionaries and trying to fit all the smaller pieces together into the larger script. But, I learned a ton by doing this project and can't wait to start the next one!