reduction update
I watched a video that said scripts shouldn't be too long so I tried to see how much I could shorten mine. Also, after looking at my works in progress, I realized I lost the functionality of displaying the player with the guardian, because not all of the kids' last names are the same as the guardians' so I changed it, even though I had to do away with all the guardians being in one list separated by commas such as the player list. For real functionality, I feel this would be a necessary change.